+++ OpenTissue/dynamics/mbd/mbd_compute_kinetic_energy_vector.h
@@ -32,7 +32,7 @@ namespace OpenTissue
     {
       typedef typename indirect_body_iterator::value_type     body_type;
       typedef typename body_type::math_policy                 math_policy;
-      typedef typename vector_type::size_type                 size_type;
+      typedef typename math_policy::size_type                 size_type;
 
       size_type n = std::distance(begin,end);