+++ OpenTissue/dynamics/mbd/mbd_get_external_force_vector.h
@@ -39,7 +39,7 @@ namespace OpenTissue
       typedef typename body_type::vector3_type                vector3_type;
       typedef typename body_type::matrix3x3_type              matrix3x3_type;
       typedef typename body_type::quaternion_type             quaternion_type;
-      typedef typename vector_type::size_type                 size_type;
+      typedef typename math_policy::size_type                 size_type;
 
       vector3_type velocity_forces;
       vector3_type force;