+++ OpenTissue/dynamics/mbd/mbd_get_factors_vector.h
@@ -51,7 +51,7 @@ namespace OpenTissue
         typedef typename group_type::math_policy                                math_policy;
         typedef typename group_type::const_indirect_constraint_iterator         const_indirect_constraint_iterator;
         typedef typename group_type::const_indirect_contact_iterator            const_indirect_contact_iterator;
-        typedef typename vector_type::size_type                                 size_type;
+        typedef typename math_policy::size_type                                 size_type;
         typedef typename math_policy::vector_range                              vector_range;
 
         math_policy::resize( mu, m);