+++ OpenTissue/dynamics/mbd/mbd_get_position_vector.h
@@ -35,7 +35,7 @@ namespace OpenTissue
       typedef typename body_type::math_policy               math_policy;
       typedef typename body_type::vector3_type              vector3_type;
       typedef typename body_type::quaternion_type           quaternion_type;
-      typedef typename vector_type::size_type               size_type;
+      typedef typename math_policy::size_type               size_type;
 
       vector3_type r;
       quaternion_type Q;